# cli/fix: use tree sitter query