--- a/app/capabilities/BUILD
+++ b/app/capabilities/BUILD
@@ -5,8 +5,5 @@ package(default_visibility = ["//visibility:public"])
 ts_library(
     name = "capabilities",
     srcs = ["capabilities.tsx"],
-    deps = [
-        "//proto:config_ts_proto",
-        "@npm//tslib",
-    ],
+    deps = ["//proto:config_ts_proto"],
 )