--- a/cli/fix/typescript/typescript.go
+++ b/cli/fix/typescript/typescript.go
@@ -1,12 +1,13 @@
 package typescript
 
 import (
+	"context"
+	_ "embed"
 	"encoding/json"
 	"flag"
-	"io/ioutil"
 	"log"
+	"os"
 	"path"
-	"regexp"
 	"sort"
 	"strings"
 
@@ -35,18 +36,26 @@ const (
 	tsxFileExtension    = ".tsx"
 )
 
-var dynamicImportPattern = regexp.MustCompile("(?ms)import\\s*\\(\\s*(['\"`]([^'\"`]+)['\"`])\\s*\\)")
-var tslibPattern = regexp.MustCompile(`(async|await|\.\.\.|import \* as)`)
+//go:embed typescript.scm
+var analysisQuery []byte
 
 type TS struct {
-	parser *sitter.Parser
+	parser  *sitter.Parser
+	tsQuery *sitter.Query
 }
 
 func NewLanguage() language.Language {
 	parser := sitter.NewParser()
 	parser.SetLanguage(tsx.GetLanguage())
+
+	q, err := sitter.NewQuery(analysisQuery, tsx.GetLanguage())
+	if err != nil {
+		log.Fatalf("Error loading query %v", err)
+	}
+
 	return &TS{
-		parser: parser,
+		parser:  parser,
+		tsQuery: q,
 	}
 }
 
@@ -79,7 +88,7 @@ func (t *TS) Kinds() map[string]rule.KindInfo {
 // files.
 func (t *TS) Loads() []rule.LoadInfo {
 	return []rule.LoadInfo{
-		rule.LoadInfo{
+		{
 			Name:    tsProjectImportPath,
 			Symbols: []string{tsProjectRuleName},
 		},
@@ -100,38 +109,60 @@ func (t *TS) Loads() []rule.LoadInfo {
 // Any non-fatal errors this function encounters should be logged using
 // log.Print.
 func (t *TS) GenerateRules(args language.GenerateArgs) language.GenerateResult {
-	rules := make([]*rule.Rule, 0)
-	imports := make([]interface{}, 0)
+	rules := make([]*rule.Rule, 0, len(args.RegularFiles))
+	imports := make([]interface{}, 0, len(args.RegularFiles))
 
 	for _, baseName := range args.RegularFiles {
 		if !strings.HasSuffix(baseName, tsFileExtension) && !strings.HasSuffix(baseName, tsxFileExtension) {
 			continue
 		}
+
 		r := rule.NewRule(tsProjectRuleName, strings.Split(baseName, ".")[0])
 		r.SetAttr(srcAttribute, []string{baseName})
 		rules = append(rules, r)
-		filePath := path.Join(args.Dir, baseName)
 
-		data, err := ioutil.ReadFile(filePath)
+		filePath := path.Join(args.Dir, baseName)
+		data, err := os.ReadFile(filePath)
 		if err != nil {
-			log.Fatalf("Error reading %s: %v", filePath, err)
+			log.Printf("error reading %s: %v\n", filePath, err)
+			continue
 		}
-		ruleImports := make([]string, 0)
-		tree := t.parser.Parse(nil, data)
-		for i := 0; i < int(tree.RootNode().ChildCount()); i++ {
-			child := tree.RootNode().Child(i)
-			if child.Type() == "import_statement" {
-				ruleImports = append(ruleImports, child.NamedChild(1).Child(1).Content(data))
+		tree, err := t.parser.ParseCtx(context.Background(), nil, data)
+		if err != nil {
+			log.Printf("error parsing %s: %v\n", filePath, err)
+			continue
+		}
+
+		qc := sitter.NewQueryCursor()
+		qc.Exec(t.tsQuery, tree.RootNode())
+
+		var ruleImports []string
+		needTSLib := false
+		for {
+			// TODO(sluongng): Check why the index return by NextCapture could not be feed into CaptureNameForId
+			//   https://github.com/smacker/go-tree-sitter/issues/97
+			cap, _, ok := qc.NextCapture()
+			if !ok {
+				break
+			}
+
+			name := t.tsQuery.CaptureNameForId(cap.Captures[0].Index)
+			switch name {
+			case "deps", "dynamic-deps":
+				for _, c := range cap.Captures {
+					ruleImports = append(ruleImports, c.Node.Content(data))
+				}
+			case "need-tslib":
+				needTSLib = true
+			default:
+				log.Printf("error unexpected capture name %s", name)
+				continue
 			}
 		}
-
-		// TODO(siggisim): See if we can grab dynamic imports and tslib usage using treesitter in an efficient way.
-		for _, match := range dynamicImportPattern.FindAllSubmatch(data, -1) {
-			ruleImports = append(ruleImports, string(match[2]))
-		}
-		if tslibPattern.Match(data) {
+		if needTSLib {
 			ruleImports = append(ruleImports, tslibImport)
 		}
+
 		imports = append(imports, ruleImports)
 	}
 
@@ -146,7 +177,6 @@ func (t *TS) GenerateRules(args language.GenerateArgs) language.GenerateResult {
 // called before the file is indexed. Unless c.ShouldFix is true, fixes
 // that delete or rename rules should not be performed.
 func (t *TS) Fix(c *config.Config, f *rule.File) {
-
 }
 
 // Resolver
@@ -237,7 +267,6 @@ func isNPMImport(importString string) bool {
 // starts. RegisterFlags may set an initial values in Config.Exts. When flags
 // are set, they should modify these values.
 func (t *TS) RegisterFlags(fs *flag.FlagSet, cmd string, c *config.Config) {
-
 }
 
 // CheckFlags validates the configuration after command line flags are parsed.
@@ -271,7 +300,7 @@ func (t *TS) Configure(c *config.Config, rel string, f *rule.File) {
 	}
 	tsConfig := c.Exts[languageName].(tsConfig)
 	packagePath := path.Join(c.RepoRoot, rel, packageFileName)
-	data, err := ioutil.ReadFile(packagePath)
+	data, err := os.ReadFile(packagePath)
 	if err != nil {
 		return
 	}