--- a/enterprise/app/code/BUILD
+++ b/enterprise/app/code/BUILD
@@ -47,7 +47,6 @@ ts_library(
         "@npm//@types/react",
         "@npm//lucide-react",
         "@npm//react",
-        "@npm//tslib",
     ],
 )