--- a/enterprise/app/history/BUILD
+++ b/enterprise/app/history/BUILD
@@ -31,7 +31,6 @@ ts_library(
         "@npm//lucide-react",
         "@npm//react",
         "@npm//rxjs",
-        "@npm//tslib",
     ],
 )
 
@@ -46,7 +45,6 @@ ts_library(
         "@npm//@types/react",
         "@npm//lucide-react",
         "@npm//react",
-        "@npm//tslib",
     ],
 )