Experimental switching to use Tree Sitter Query to collect Typescript metadata.

This comes with a performance tradeoff:

- Before this PR
  ```
  > hyperfine 'bb fix'
  Benchmark 1: bb fix
    Time (mean ± σ):      1.712 s ±  0.007 s    [User: 1.732 s, System: 0.091 s]
    Range (min … max):    1.703 s …  1.724 s    10 runs
  ```

- After this PR
  ```
  > hyperfine 'bb fix'
  Benchmark 1: bb fix
    Time (mean ± σ):      2.432 s ±  0.015 s    [User: 2.435 s, System: 0.100 s]
    Range (min … max):    2.419 s …  2.471 s    10 runs
  ```

So it's `1.732 -> 2.435 (+40.5%)` slower.

But we gain the ability to precisely finding structure within the Typescript AST.

This PR is split into 3 commits for easier reviewing:
```
> PAGER= git log -3 --oneline --reverse
# Apply 'bb fix' to latest master
5991b4f14 apply bb fix

# Apply new code changes to switch to TreeSitter Query 
77b711b02 cli/fix: switch to use tree-sitter

# Apply new 'bb fix' with new metadata extracted using TreeSitter Query
adde3f78b apply bb fix
```

---

**version bump**: patch