@sluongng 2023-03-15
I cook this up during my downtime prep-ing for flights / during flights / post-flight-but-no-5G.

Even though it's slower than regex approach, now we have the ability to generate more intricate rules (i.e. the future rules_js / rules_ts from Aspect folks).

This could also be applied to other languages in the tree-sitter ecosystem, including https://github.com/amaanq/tree-sitter-starlark