@siggisim 2023-03-15
My initial version used tree-sitter queries like this, but I found that it was much slower than I liked. I want `bb fix` to be fast so it can be run regularly (i.e. on every save). 

Our repo only has 157 .ts/.tsx files, so I'm a bit worried on how this will scale to larger repos. 

I'd almost rather go back to the regex approach than slow things down considerably. There might be some other solution like only running this on changed files that would make this good enough though.