@sluongng 2023-03-23
Yeah I think changed files might be an interesting direction to take the perf.
Detecting file changes is tricky though, https://lobste.rs/s/mjppkn/navigating_complexities_file_watching is a good thread on the topic. I know that we are using github.com/bduffany/godemon which in turn uses github.com/fsnotify/fsnotify. I will do a bit more research and come back to this in the future.